+++ client/src/components/Topbar/MenuInformation.vue
@@ -73,6 +73,8 @@
     <a
       :href="$url('guide')"
       class="dropdown-item sub"
+      target="_blank"
+      rel="noopener noreferrer nofollow"
       role="menuitem"
     >
       {{ $i18n('menu.entry.guide') }}