+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Removed quiz description #737 !1565 @chriswalg
 - Removed unused legal agreement for new amb's #821 !1619 @chriswalg
 - Workgroup Admins of Startpage and Team/Partner page can edit their pages #967 !1651 @fs_k
+- Moved guide page (ratgeber) to wiki page #776 !1567 @chriswalg
 
 ## Bugfixes
 - Filter outdated baskets from REST responses #706 !1608 @alex.simm